shuffle

+++ MetalNeuron/src/data-source-manager.cpp
@@ -11,8 +11,8 @@ DataSourceManager::DataSourceManager(int inputDim, int hiddenDim, int outputDim,
 DataSourceManager::~DataSourceManager() {}
 
 void DataSourceManager::initialize(std::function<void()> onComplete,
-                                   double (*inputFunc)(double, int),
-                                   double (*targetFunc)(double, int))
+                                   double (*inputFunc)(double, double),
+                                   double (*targetFunc)(double, double))
 {
     for (int t = 0; t < sequenceLength_; ++t) {
         x.buildAsyncAtTimestep(inputFunc, t, [this, t, targetFunc, onComplete]() {
@@ -28,7 +28,7 @@ void DataSourceManager::initialize(std::function<void()> onComplete,
     }
 }
 
-void DataSourceManager::buildInputAtTimestep(std::function<double(double, int)> inputFunc, int timestep, std::function<void()> onComplete) {
+void DataSourceManager::buildInputAtTimestep(std::function<double(double, double)> inputFunc, int timestep, std::function<void()> onComplete) {
     x.buildAsyncAtTimestep(inputFunc, timestep, [onComplete]() {
         dispatch_async(dispatch_get_main_queue(), ^{
             onComplete();
@@ -36,7 +36,7 @@ void DataSourceManager::buildInputAtTimestep(std::function<double(double, int)> 
     });
 }
 
-void DataSourceManager::buildTargetAtTimestep(std::function<double(double, int)> targetFunc, int timestep, std::function<void()> onComplete) {
+void DataSourceManager::buildTargetAtTimestep(std::function<double(double, double)> targetFunc, int timestep, std::function<void()> onComplete) {
     y_hat.buildAsyncAtTimestep(targetFunc, timestep, [onComplete]() {
         dispatch_async(dispatch_get_main_queue(), ^{
             onComplete();
